--- accounts/auth/__init__.py
@@ -1,2 +1,0 @@
-from .basic import BasicAuth
-from .jwt import JWTAuth
